Added a few more unit tests on the NNetwork class and added a few more load methods.

+++ src/main/scala/com/nn/utility/DataLoader.scala
@@ -4,20 +4,21 @@ import scala.collection.mutable.ArrayBuffer
 import scala.io.BufferedSource
 
 /**
+ * Static Object to load general purpose csv data.
  * Created by george on 12/6/14.
  */
 object DataLoader {
 
   /**
-   * Input the training/testing data to be read by the network. This function currently assumes the data is a csv
+   * Input the training/testing data to be read by the network. This function currently assumes the data is a comma delimited csv file
    */
-  def loadCsvData(data: BufferedSource, delim: String): ArrayBuffer[Array[Int]] = {
+  def loadCsvData(data: BufferedSource, delim: String): ArrayBuffer[Array[Double]] = {
     val iter = data.getLines().drop(1).map(_.split(delim))
-    var dataSet = new ArrayBuffer[Array[Int]]()
+    var dataSet = new ArrayBuffer[Array[Double]]()
     // read in each example and convert to integer arrays
     while(iter.hasNext){
       try {
-        val example: Array[Int] = iter.next().map(_.toInt)
+        val example: Array[Double] = iter.next().map(_.toDouble)
         dataSet.+=(example)
       }catch{
         case numFE: NumberFormatException => {
